Split LLK pack APIs by operation family.
