+++ tt_metal/hw/ckernels/blackhole/metal/llk_api/experimental/llk_pack_fast_tilize_api.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "llk_pack_api.h"
+#include "llk_pack_common_api.h"
 #include "experimental/llk_pack_fast_tilize.h"
 
 /*************************************************************************
