+++ tt_metal/tt-llk/tests/helpers/include/llk_lib_pack_wrappers.h
@@ -11,6 +11,7 @@
 #include <cstdint>
 
 #include "llk_pack.h"
+#include "llk_pack_untilize.h"
 
 using ckernel::FACE_R_DIM;
 using ckernel::TILE_C_DIM;
