+++ tt_metal/tt-llk/tests/sources/pack_untilize_test.cpp
@@ -113,6 +113,7 @@ void run_kernel(RUNTIME_PARAMETERS params)
 
 #include "llk_lib_pack_wrappers.h"
 #include "llk_pack_common.h"
+#include "llk_pack_untilize.h"
 #include "params.h"
 
 constexpr std::uint32_t L1_ACCESS_ADDRESS_GRANULARITY = 16; // in bytes
